fix modules linking
