+++ api/src/helpers/application.rs
@@ -1,5 +1,6 @@
 use crate::auth::user::User as AuthUser;
 use crate::errors::*;
+use crate::server::AppState;
 use actix_web::{http, http::StatusCode, Body::Binary, HttpResponse, Responder};
 use serde_json::{self, Value};
 use std::collections::HashMap;
