+++ config/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "bigneon_config"
+version = "0.1.0"
+authors = ["Big Neon Team"]
+edition = "2018"
+description = "Big Neon application configuration"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bigneon_errors = { version = "0.1.0", path = "../errors" }
+bigneon_db = { path = "../db" }
+dotenv = "0.13"
+itertools = "0.7"
+tari-client= {path="../tari-client"}
+serde = { version = "1", default_features = false, features = ["derive"] }
