+++ config/src/lib.rs
@@ -1,0 +1,3 @@
+mod config;
+
+pub use config::*;
