+++ db_connections/src/readonly_connection.rs
@@ -1,10 +1,5 @@
-use crate::db::*;
-use crate::errors::BigNeonError;
-use crate::server::AppState;
-use actix_web::{FromRequest, HttpRequest, Result};
+use crate::*;
 use diesel;
-use diesel::connection::TransactionManager;
-use diesel::Connection as DieselConnection;
 use diesel::PgConnection;
 use std::sync::Arc;
 
@@ -68,27 +63,5 @@ impl Clone for ReadonlyConnection {
         ReadonlyConnection {
             inner: self.inner.clone(),
         }
-    }
-}
-
-impl FromRequest<AppState> for ReadonlyConnection {
-    type Config = ();
-    type Result = Result<ReadonlyConnection, BigNeonError>;
-
-    fn from_request(request: &HttpRequest<AppState>, _config: &Self::Config) -> Self::Result {
-        if let Some(connection) = request.extensions().get::<ReadonlyConnection>() {
-            return Ok(connection.clone());
-        }
-
-        let connection = request.state().database_ro.get_ro_connection()?;
-        {
-            let connection_object = connection.get();
-            connection_object
-                .transaction_manager()
-                .begin_transaction(connection_object)?;
-        }
-
-        request.extensions_mut().insert(connection.clone());
-        Ok(connection)
     }
 }
