+++ domain_events/src/executors/process_settlement_report.rs
@@ -1,11 +1,13 @@
-use crate::config::Config;
-use crate::db::Connection;
-use crate::domain_events::executor_future::ExecutorFuture;
-use crate::domain_events::routing::DomainActionExecutor;
-use crate::errors::*;
+use crate::executor_future::ExecutorFuture;
+use crate::routing::DomainActionExecutor;
+use bigneon_config::Config;
+use bigneon_db_connections::Connection;
+use bigneon_errors::*;
 use bigneon_db::prelude::*;
 use futures::future;
 use log::Level::Error;
+use serde_json::json;
+use logging::jlog;
 
 pub struct ProcessSettlementReportExecutor {
     config: Config,
