+++ domain_events/src/executors/send_order_complete.rs
@@ -1,14 +1,16 @@
-use crate::communications::mailers;
-use crate::config::Config;
-use crate::db::Connection;
-use crate::domain_events::executor_future::ExecutorFuture;
-use crate::domain_events::routing::DomainActionExecutor;
-use crate::errors::*;
+use crate::executor_future::ExecutorFuture;
+use crate::routing::DomainActionExecutor;
+use bigneon_config::Config;
+use bigneon_db_connections::Connection;
+use bigneon_errors::*;
+use bigneon_communications::mailers;
 use bigneon_db::prelude::*;
 use futures::future;
 use log::Level::Error;
 use std::collections::HashMap;
 use uuid::Uuid;
+use serde_json::json;
+use logging::jlog;
 
 pub struct SendOrderCompleteExecutor {
     config: Config,
