+++ domain_events/src/executors/submit_sitemap_to_search_engines.rs
@@ -1,11 +1,13 @@
-use crate::db::Connection;
-use crate::domain_events::executor_future::ExecutorFuture;
-use crate::domain_events::routing::DomainActionExecutor;
-use crate::errors::BigNeonError;
+use crate::executor_future::ExecutorFuture;
+use crate::routing::DomainActionExecutor;
+use bigneon_db_connections::Connection;
+use bigneon_errors::BigNeonError;
 use bigneon_db::prelude::*;
 use futures::future;
 use log::Level;
 use reqwest::Client;
+use serde_json::json;
+use logging::jlog;
 
 pub struct SubmitSitemapToSearchEnginesExecutor {
     api_url: String,
