+++ domain_events/src/executors/update_genres.rs
@@ -1,11 +1,14 @@
-use crate::db::Connection;
-use crate::domain_events::executor_future::ExecutorFuture;
-use crate::domain_events::routing::DomainActionExecutor;
-use crate::errors::*;
+use crate::executor_future::ExecutorFuture;
+use crate::routing::DomainActionExecutor;
+use bigneon_db_connections::Connection;
+use bigneon_errors::*;
 use bigneon_db::prelude::*;
 use futures::future;
 use log::Level::Error;
 use uuid::Uuid;
+use serde::{Serialize, Deserialize};
+use serde_json::json;
+use logging::jlog;
 
 pub struct UpdateGenresExecutor {}
