+++ locators/src/lib.rs
@@ -1,0 +1,5 @@
+mod deep_linker;
+mod service_locator;
+
+pub use deep_linker::*;
+pub use service_locator::*;
