+++ payments/src/payment_processor.rs
@@ -1,7 +1,8 @@
-use crate::payments::*;
+use crate::*;
 use bigneon_db::models::PaymentProviders;
 use chrono::NaiveDateTime;
 use uuid::Uuid;
+use serde::Serialize;
 
 pub enum PaymentProcessorBehavior {
     AuthThenComplete(Box<dyn AuthThenCompletePaymentBehavior>),
