+++ payments/src/stripe/mod.rs
@@ -1,7 +1,7 @@
-use crate::payments::*;
+use crate::*;
 use bigneon_db::models::PaymentProviders;
-use stripe::StripeClient;
-use stripe::StripeError;
+use ::stripe::StripeClient;
+use ::stripe::StripeError;
 
 impl From<StripeError> for PaymentProcessorError {
     fn from(s: StripeError) -> PaymentProcessorError {
